ref(slicing): Minor refactoring for slicing
Rename partitioning to slicing in order to be consistent between
sentry and snuba. The metrics indexer only writes to a single topic
for different metric types, so define a single sliceable entity.
